requirements.txt
pandas
requests
yfinance
plotly
python-dotenv
